e2fsprogs: Prepare for adding and using static libs
 * Convert misc utils to inline.  This needed to avoid duplicate
   symbols when linking multiple e2fsprogs libs into an executable.

 * Ensure macros are set properly in blkid_types.h.  __bitwise and
   __force are defined in uapi headers, so we must ensure they are
   undefined in order to build against code that uses any libc or
   kernel headers.

 * Add BUILD_AS_LIB guards to main() and program_name where needed,
   following the existing style in misc/tune2fs.c.
