+++ misc/Android.bp
@@ -37,7 +37,6 @@ cc_binary {
 
     srcs: [
         "mke2fs.c",
-        "util.c",
         "mk_hugefiles.c",
         "default_profile.c",
     ],
@@ -108,7 +107,6 @@ cc_defaults {
     defaults: ["e2fsprogs-defaults"],
     srcs: [
         "tune2fs.c",
-        "util.c",
     ],
     cflags: ["-DNO_RECOVERY"],
     include_dirs: ["external/e2fsprogs/e2fsck"],
