+++ misc/tune2fs.c
@@ -72,7 +72,11 @@ extern int optind;
 
 extern int ask_yn(const char *string, int def);
 
+#ifndef BUILD_AS_LIB
 const char *program_name = "tune2fs";
+#else
+extern const char *program_name;
+#endif
 char *device_name;
 char *new_label, *new_last_mounted, *new_UUID;
 char *io_options;
